% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css_generator.R
\name{css_generator}
\alias{css_generator}
\title{css_generator}
\usage{
css_generator(path_logo = NULL)
}
\arguments{
\item{path_logo}{The path to logo, will automatically be guessed on the package.}
}
\value{
A css string
}
\description{
Generate css to be included in quarto.
}
