% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_round.R
\name{custom_round}
\alias{custom_round}
\title{Custom Round and Format}
\usage{
custom_round(x, digits = 1)
}
\arguments{
\item{x}{A numeric vector to be rounded and formatted.}

\item{digits}{An integer indicating the number of decimal places to round to. Defaults to 1.}
}
\value{
A character vector with the rounded and formatted numbers.
}
\description{
Rounds a numeric value to a specified number of decimal places and formats it
to always show the specified number of decimal places, including trailing zeros.
}
\examples{
custom_round(3.14159)      # "3.1"
custom_round(3.14159, 3)   # "3.142"
custom_round(c(2, 2.5), 2) # "2.00" "2.50"

}
