% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_degroup.R
\name{desc_degroup}
\alias{desc_degroup}
\title{Description of grouped and ungrouped tables with data management}
\usage{
desc_degroup(
  data1,
  table_title = "",
  quali = NULL,
  quanti = NULL,
  round_quanti = c(1),
  round_quali = c(0, 1),
  var_title = "Variable",
  var_group = NULL,
  group_title = ""
)
}
\arguments{
\item{data1}{Dataframe. The input data to generate the table.}

\item{table_title}{Character. Title of the table (default: "").}

\item{quali}{Character vector. A vector containing the names of the qualitative
variables to describe (default: NULL).}

\item{quanti}{Character vector. A vector containing the names of the quantitative
variables to describe (default: NULL).}

\item{round_quanti}{Numeric vector. Number of decimal places for quantitative
variables (default: 1).}

\item{round_quali}{Numeric vector. Number of decimal places for qualitative variables
(default: c(0, 1)).}

\item{var_title}{Character. Title of the variable column in the table (default: "Variable").}

\item{var_group}{Character. Name of the variable used to group the data (default: NULL).}

\item{group_title}{Character. Title for the group variable (default: "").}
}
\value{
A \code{gtsummary} object representing the descriptive table.
}
\description{
This function generates a statistical summary of grouped and ungrouped data
using the \code{gtsummary} package. It allows displaying customized statistics for
categorical and continuous variables, with options to adjust titles, rounding,
and other parameters.
}
\details{
\itemize{
\item Categorical variables are displayed with statistics like \verb{\{n\} (\{p\}\%)}.
\item Continuous variables are displayed with statistics such as
\code{{mean} ({sd})}, \verb{\{min\} ; \{max\}}, \verb{\{median\} (\{p25\};\{p75\})}.
\item Columns are formatted and customized with options for style, titles, and footnotes.
}
}
\examples{
# Example usage:
data <- data.frame(
  group = c("A", "B", "A", "B"),
  gender = factor(c("Male", "Female", "Female", "Male")),
  age = c(34, 29, 45, 36)
)
desc_degroup(
  data1 = data,
  table_title = "Descriptive Table",
  quali = c("gender"),
  quanti = c("age"),
  var_group = "group"
)

}
