% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_degroup_group.R
\name{desc_degroup_group}
\alias{desc_degroup_group}
\title{Description of grouped tables with data management}
\usage{
desc_degroup_group(
  data1,
  table_title = "",
  quali = NULL,
  quanti = NULL,
  round_quanti = c(1),
  round_quali = c(0, 1),
  var_title = "Variable",
  var_group = NULL,
  group_title = "",
  tests = FALSE
)
}
\arguments{
\item{data1}{Dataframe. The input data to generate the table.}

\item{table_title}{Character. Title of the table (default: "").}

\item{quali}{Character vector. A vector containing the names of qualitative
variables to describe (default: NULL).}

\item{quanti}{Character vector. A vector containing the names of quantitative
variables to describe (default: NULL).}

\item{round_quanti}{Numeric vector. Number of decimal places for quantitative
variables (default: 1).}

\item{round_quali}{Numeric vector. Number of decimal places for qualitative variables
(default: c(0, 1)).}

\item{var_title}{Character. Title of the variable column in the table (default: "Variable").}

\item{var_group}{Character. Name of the variable used to group the data (default: NULL).}

\item{group_title}{Character. Title for the group variable (default: "").}

\item{tests}{A value in order to add p value. Defaut to \code{FALSE} OPTION :
\itemize{
\item \code{FALSE}: No p-value add
\item \code{TRUE}: Add p-value made by default by gtsummary. See gtsummary add_p() options.
\item \code{list()}: To force tests. See gtsummary add_p() options.
}}
}
\value{
A \code{gtsummary} object representing the descriptive table for grouped data.
}
\description{
This function generates a statistical summary for grouped data using the
\code{gtsummary} package. It provides customized statistics for both categorical
and continuous variables, allowing for adjustments in titles, rounding, and
other formatting options.
}
\details{
\itemize{
\item Categorical variables are displayed with statistics like \verb{\{n\} (\{p\}\%)}.
\item Continuous variables are displayed with statistics such as
\code{{mean} ({sd})}, \verb{\{median\} (\{p25\}; \{p75\})}, and \verb{\{min\} ; \{max\}}.
\item Columns and rows are formatted and customized with options for titles,
rounding, footnotes, and group titles.
\item The \code{by} argument is used to create grouped tables based on a specified grouping variable.
}
}
\examples{
# Example usage:
data <- data.frame(
  group = c("A", "B", "A", "B"),
  gender = factor(c("Male", "Female", "Female", "Male")),
  age = c(34, 29, 45, 36)
)
desc_degroup_group(
  data1 = data,
  table_title = "Descriptive Table",
  quali = c("gender"),
  quanti = c("age"),
  var_group = "group"
)

}
