% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_ei_per_grade.R
\name{desc_ei_per_grade_df_to_gt}
\alias{desc_ei_per_grade_df_to_gt}
\title{desc_ei_per_grade_df_to_gt}
\usage{
desc_ei_per_grade_df_to_gt(df_wide, vec_grp, language = "fr")
}
\arguments{
\item{df_wide}{A wide-format dataframe summarizing AE counts and percentages by grade and group.}

\item{vec_grp}{A vector of unique group names.}

\item{language}{'fr' default or 'en'}
}
\value{
A formatted gt table.
}
\description{
Converts the processed AE grade dataframe into a gt table for visualization.
}
\keyword{internal}
