% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_ei_per_grade.R
\name{desc_ei_per_grade_prepare_df}
\alias{desc_ei_per_grade_prepare_df}
\title{desc_ei_per_grade_prepare_df}
\usage{
desc_ei_per_grade_prepare_df(
  augmented_df_pat_grp,
  augmented_df_pat_grade_grp,
  digits = 1
)
}
\arguments{
\item{augmented_df_pat_grp}{A dataframe with patient IDs and groups, including a "Any grade" group.}

\item{augmented_df_pat_grade_grp}{A dataframe with patient IDs, grades, and groups.}

\item{digits}{Number of digits for percentages}
}
\value{
A dataframe in wide format with AE counts and percentages by grade and group.
}
\description{
Prepares a wide-format dataframe summarizing AE by grade and group.
}
\keyword{internal}
