% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_ei_per_pt.R
\name{desc_ei_per_pt_df_to_gt}
\alias{desc_ei_per_pt_df_to_gt}
\title{Convert AE Summary Data to GT Table}
\usage{
desc_ei_per_pt_df_to_gt(df_wide, vec_grp, language = "fr")
}
\arguments{
\item{df_wide}{A wide-format dataframe containing summarized AE data.}

\item{vec_grp}{A character vector of group names for which AE data is presented.}

\item{language}{'fr' default or 'en'}
}
\value{
A \code{gt} table formatted with appropriate labels, spans, and styling.
}
\description{
This function takes a prepared wide-format dataframe summarizing adverse events and patients
and converts it into a formatted \code{gt} table for easy visualization.
}
\keyword{internal}
