% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_ei_per_soc.R
\name{desc_ei_per_soc}
\alias{desc_ei_per_soc}
\title{desc_ei_per_soc}
\usage{
desc_ei_per_soc(df_soc_pt, df_pat_grp, df_pat_soc)
}
\arguments{
\item{df_soc_pt}{A dataframe with two columns: pt and soc}

\item{df_pat_grp}{A dataframe with two columns: id_pat and grp (the rct arm)}

\item{df_pat_soc}{A dataframe with two columns: id_pat and soc}
}
\value{
A gt table
}
\description{
A function to describe AE by soc and pt
}
\examples{
df_soc_pt <- data.frame(
  soc = c("Arrhythmia",
          "Myocardial Infarction",
          "Pneumonia",
          "Sepsis"),
  pt = c("Cardiac Disorders",
         "Cardiac Disorders",
         "Infections",
         "Infections")
)

df_pat_grp <- data.frame(id_pat = paste0("ID_", 1:10),
                         grp = c(rep("A", 3), rep("B", 3), rep("C", 4)))

df_pat_soc <- data.frame(id_pat = c("ID_1", "ID_1",
                                    "ID_2",
                                    "ID_4",
                                    "ID_9"),
                         soc = c("Arrhythmia", "Myocardial Infarction",
                                 "Arrhythmia", "Pneumonia",
                                 "Pneumonia"))

desc_ei_per_soc(df_soc_pt = df_soc_pt,
                df_pat_grp = df_pat_grp,
                df_pat_soc = df_pat_soc)

}
