% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_ei_per_soc.R
\name{desc_ei_per_soc_prepare_df}
\alias{desc_ei_per_soc_prepare_df}
\title{Prepare Data for AE Description by SOC and PT}
\usage{
desc_ei_per_soc_prepare_df(augmented_df_pat_grp, augmented_df_pat_soc_grp)
}
\arguments{
\item{augmented_df_pat_grp}{A dataframe containing patient IDs and group assignments, including a "Total" group.}

\item{augmented_df_pat_soc_grp}{A dataframe linking patient IDs to SOC and PT, with group assignments.}
}
\value{
A wide-format dataframe summarizing adverse event occurrences and patient counts across groups.
}
\description{
This function processes patient and adverse event data to generate a wide-format summary dataframe,
including total counts and percentages of events and patients per SOC (System Organ Class) and PT (Preferred Term).
}
\examples{
\dontrun{
df_wide <- desc_ei_per_soc_prepare_df(augmented_df_pat_grp, augmented_df_pat_soc_grp)
}
}
\keyword{internal}
