% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_group.R
\name{desc_group}
\alias{desc_group}
\title{Description of summary tables with data management}
\usage{
desc_group(
  data1,
  table_title = "",
  quali = NULL,
  quanti = NULL,
  round_quanti = c(1),
  round_quali = c(0, 1),
  var_title = "Variable"
)
}
\arguments{
\item{data1}{Dataframe. The input data to generate the summary table.}

\item{table_title}{Character. Title of the table (default: "").}

\item{quali}{Character vector. A vector containing the names of the qualitative
variables to describe (default: NULL).}

\item{quanti}{Character vector. A vector containing the names of the quantitative
variables to describe (default: NULL).}

\item{round_quanti}{Numeric vector. Number of decimal places for quantitative
variables (default: 1).}

\item{round_quali}{Numeric vector. Number of decimal places for qualitative variables
(default: c(0, 1)).}

\item{var_title}{Character. Title of the variable column in the table (default: "Variable").}
}
\value{
A \code{gtsummary} object representing the summary table.
}
\description{
This function generates a statistical summary table for both categorical
and continuous variables using the \code{gtsummary} package. It formats and displays
statistics for the given variables, allowing for customized rounding and other
formatting options.
}
\details{
\itemize{
\item Categorical variables are displayed with statistics such as \verb{\{n\} (\{p\}\%)}.
\item Continuous variables are displayed with statistics such as \code{{mean} ({sd})},
\verb{\{min\} ; \{max\}}, and \verb{\{median\} (\{p25\}; \{p75\})}.
\item The table is formatted with bold labels for variables and italicized levels for
the categories.
\item Missing data is excluded from the summary (\code{missing = "no"}).
\item The total column is labeled as "\strong{TOTAL} (n={N})".
\item The footnotes are removed from the table using \code{modify_footnote}.
}
}
\examples{
# Example usage:
data <- data.frame(
  group = c("A", "B", "A", "B"),
  gender = factor(c("Male", "Female", "Female", "Male")),
  age = c(34, 29, 45, 36)
)
desc_group(
  data1 = data,
  table_title = "Descriptive Table",
  quali = c("gender"),
  quanti = c("age"),
  var_title = "Variable"
)

}
