% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_builder_ae.R
\name{df_builder_ae}
\alias{df_builder_ae}
\title{Prepare a dataframe for creating AE plots}
\usage{
df_builder_ae(
  df_pat_grp,
  df_pat_llt,
  id_col = "USUBJID",
  group_col = "RDGRPNAME",
  ei_num_col = "EINUM",
  ei_llt_col = "EILLTN",
  ei_soc_col = "EISOCPN",
  ei_pt_col = "EIPTN",
  ref_grp = NULL
)
}
\arguments{
\item{df_pat_grp}{A dataframe with two columns: USUBJID (Patient id) and RDGRPNAME (the RCT arm).}

\item{df_pat_llt}{A dataframe with columns: USUBJID (patient id), EINUM (AE id), llt (AE LLT), pt (AE PT), soc (AE SOC)}

\item{id_col}{Patient id column (default: "USUBJID").}

\item{group_col}{group column, the rct arm (default: "RDGRPNAME").}

\item{ei_num_col}{AE id column (default: "EINUM").}

\item{ei_llt_col}{AE LLT column (default: "EILLTN").}

\item{ei_soc_col}{AE SOC column (default: "EISOCPN").}

\item{ei_pt_col}{AE PT column (default: "EIPTN")}

\item{ref_grp}{(Optional) A reference group for comparisons. Defaults to the first group in \code{df_pat_grp}.}
}
\value{
A dataframe with all the info to build AE plots
}
\description{
Prepare a dataframe for creating AE plots
}
