% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_name_to_adress.R
\name{from_name_to_adress}
\alias{from_name_to_adress}
\title{Convert a Name to an Email Address}
\usage{
from_name_to_adress(name = "Peter Parker")
}
\arguments{
\item{name}{A character string representing a full name. Default is "Peter Parker".}
}
\value{
A character string containing the generated email address.
}
\description{
This function transforms a given name into an email address following the format \code{firstname.lastname@chu-bordeaux.fr}.
}
\examples{
from_name_to_adress("John Doe")
from_name_to_adress()

}
