% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indent_table.R
\name{indent_table}
\alias{indent_table}
\title{Text indentation for \code{gtsummary} tables}
\usage{
indent_table(g_table, indent = 0)
}
\arguments{
\item{g_table}{A \code{gt} table object}

\item{indent}{A numerical value corresponding to the pixel value, which defines the text indentation (default = 0 corresponding to px(0)). 30 ~ px(30) should be a good compromise.}
}
\value{
A \code{gt} table object with indentation applied.
}
\description{
Text indentation for \code{gtsummary} tables
}
\examples{
tbl_bis <- RastaRocket::desc_var(
iris,
table_title = "test",
quali = "Species")

tbl_bis |> indent_table(indent = 30)
tbl_bis |> indent_table(indent = 60)

}
