% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskdifference.R
\name{riskdifference}
\alias{riskdifference}
\title{riskdifference}
\usage{
riskdifference(a, b, N1, N0, CRC = FALSE, conf.level = 0.95)
}
\arguments{
\item{a}{The number of disease occurence among exposed cohort.}

\item{b}{The number of disease occurence among non-exposed cohort.}

\item{N1}{The population at risk of the exposed cohort.}

\item{N0}{The population at risk of the unexposed cohort.}

\item{CRC}{Logical.  If TRUE, calculate confidence intervals for each risk.  Default is FALSE.}

\item{conf.level}{Probability for confidence intervals.  Default is 0.95.}
}
\value{
A list with the results
}
\description{
A function from the fmsb package to compute risk difference. Calculate risk
difference (a kind of attributable risk / excess  risk) and its confidence
intervals based on approximation, followed by  null hypothesis
(risk difference equals to 0) testing.
}
