% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_css.R
\name{write_css}
\alias{write_css}
\title{Generate a CSS File}
\usage{
write_css(path)
}
\arguments{
\item{path}{Character string specifying the file path where the CSS file will be saved.}
}
\value{
None. The function writes a CSS file to the specified file path.
}
\description{
This function creates and writes a CSS file with predefined styling for tables and text formatting.
}
