% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_datestamp_output_file.R
\name{write_datestamp_output_file}
\alias{write_datestamp_output_file}
\title{write_datestamp_output_file}
\usage{
write_datestamp_output_file(output_folder, path, from_file)
}
\arguments{
\item{output_folder}{The output folder}

\item{path}{The path of the R script}

\item{from_file}{The initial html file to be renamed}
}
\value{
Nothing
}
\description{
A function to write a R file and add datestamp
}
