% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_quarto_yml.R
\name{write_quarto_yml}
\alias{write_quarto_yml}
\title{write_quarto_yml}
\usage{
write_quarto_yml(path)
}
\arguments{
\item{path}{The path toward quarto yaml file}
}
\value{
nothing
}
\description{
Write quarto extension
}
