% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rendercopy.R
\name{write_rendercopy}
\alias{write_rendercopy}
\title{write_rendercopy}
\usage{
write_rendercopy(output_folder, path, name)
}
\arguments{
\item{output_folder}{The output folder}

\item{path}{The path of the R script}

\item{name}{The name of the files}
}
\value{
Nothing
}
\description{
A function to write a R file rendercopy
}
