% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_missing_info.R
\name{add_missing_info}
\alias{add_missing_info}
\title{Add missing value information to a gtsummary table}
\usage{
add_missing_info(base_table)
}
\arguments{
\item{base_table}{A \code{gtsummary} table object.}
}
\value{
A \code{gtsummary} table object with missing value information and modifications applied.
}
\description{
This function merge missing data row into label row of gtsummary object
}
