% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_ei_per_grade.R
\name{desc_ei_per_grade}
\alias{desc_ei_per_grade}
\title{desc_ei_per_grade}
\usage{
desc_ei_per_grade(df_pat_grp, df_pat_grade)
}
\arguments{
\item{df_pat_grp}{A dataframe with two columns: id_pat and grp (the RCT arm).}

\item{df_pat_grade}{A dataframe with two columns: id_pat, num_ae (the AE id) and grade (the AE grade).}
}
\value{
A gt table summarizing the AE by grade.
}
\description{
A function to describe adverse events (AE) by grade.
}
\examples{
df_pat_grp <- data.frame(id_pat = paste0("ID_", 1:10),
                         grp = c(rep("A", 3), rep("B", 3), rep("C", 4)))

df_pat_grade <- data.frame(id_pat = c("ID_1", "ID_1",
                                      "ID_2",
                                      "ID_4",
                                      "ID_9"),
                           num_ae = c(1, 2,
                                      1,
                                      1,
                                      1),
                           grade = c(1, 3,
                                     4,
                                     2,
                                     4))

desc_ei_per_grade(df_pat_grp = df_pat_grp,
                  df_pat_grade = df_pat_grade)

}
