% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_var_function.R
\name{desc_var}
\alias{desc_var}
\title{Generate Descriptive Tables for Variables}
\usage{
desc_var(
  data1,
  table_title = "",
  quali = NULL,
  quanti = NULL,
  group,
  var_title = "Variable",
  var_group = NULL,
  group_title = "",
  round_quanti = c(1),
  round_quali = c(0, 1),
  DM = "NULL",
  tests = FALSE,
  show_n_per_group = FALSE,
  show_missing_data = NULL
)
}
\arguments{
\item{data1}{A data frame containing the dataset to be analyzed.}

\item{table_title}{A character string specifying the title of the table.}

\item{quali}{A vector of qualitative variables to be described. Defaults to \code{NULL}.}

\item{quanti}{A vector of quantitative variables to be described. Defaults to \code{NULL}.}

\item{group}{A logical value or character string indicating whether the variables should be grouped. Options:
\itemize{
\item \code{"ALL"}: Grouped by all levels of \code{var_group}.
\item \code{TRUE}: Grouped by \code{var_group}.
\item \code{FALSE}: Not grouped.
}}

\item{var_title}{A character string for the title of the variable column in the table. Defaults to \code{"Variable"}.}

\item{var_group}{A variable used for grouping (if applicable). Defaults to \code{NULL}.}

\item{group_title}{A character string specifying the title for the grouping variable.}

\item{round_quanti}{An integer vector specifying the number of decimal places for quantitative variables. Defaults to \code{c(1)}.}

\item{round_quali}{An integer vector specifying the number of decimal places for qualitative variables. Defaults to \code{c(0, 1)}.}

\item{DM}{A character string controlling the handling of missing data. Options:
\itemize{
\item \code{"tout"}: Keep all variables and drop unused levels of factors.
\item \code{"tri"}: Retain only variables with at least three levels.
\item \code{"remove"}: Remove missing levels from factors.
\item \code{"NULL"}: No special handling of missing data.
}}

\item{tests}{A value in order to add p value. Defaut to \code{FALSE} OPTION :
\itemize{
\item \code{FALSE}: No p-value add
\item \code{TRUE}: Add p-value made by default by gtsummary. See gtsummary add_p() options.
\item \code{list()}: To force tests. See gtsummary add_p() options.
}}

\item{show_n_per_group}{Default to \code{FALSE}. Should the 'N' appears in the column header of the groups. Can be either :
\itemize{
\item \code{FALSE}: No 'N' is shown
\item \code{TRUE}: 'N' is shown
}}

\item{show_missing_data}{Default to \code{NULL}. Should the missing data be displayed. Can be either :
\itemize{
\item \code{FALSE}: No missing data displayed
\item \code{TRUE}: Missing data displayed
\item \code{NULL} (default): will be switch to \code{anyNA(data1)} value.
}}
}
\value{
A gtsummary table object containing the descriptive statistics.
}
\description{
This function creates descriptive tables for variables in a dataset. It can handle qualitative and quantitative variables, grouped or ungrouped, and supports multiple configurations for handling missing data (DM).
}
\details{
The function processes the dataset according to the specified parameters and generates descriptive tables.
\itemize{
\item It first uses the \code{ajouter_label_ndm()} function to append missing data statistics to variable labels.
\item Depending on the \code{group} and \code{DM} arguments, it adjusts the dataset and creates tables using helper functions like \code{desc_group}, \code{desc_degroup}, and \code{desc_degroup_group}.
\item Qualitative variables are reordered, and unused levels are dropped when necessary.
}
}
\examples{
# Example usage:
library(dplyr)

# Sample dataset
data1 <- data.frame(
  group = c("A", "B", "A", "C"),
  var1 = c(1, 2, 3, NA),
  var2 = c("X", "Y", "X", NA)
)

# Generate descriptive table
table <- desc_var(
  data1 = data1,
  table_title = "Descriptive Table",
  group = FALSE,
  group_title = "Group",
  DM = "tout"
)

}
