% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_builder_ae.R
\name{df_builder_ae}
\alias{df_builder_ae}
\title{Prepare a dataframe for creating AE plots}
\usage{
df_builder_ae(df_pat_grp, df_pat_llt, ref_grp = NULL)
}
\arguments{
\item{df_pat_grp}{A data frame of patient groups. Must contain columns \code{USUBJID } (patient ID)
and \code{RDGRPNAME} (group assignment).}

\item{df_pat_llt}{A data frame with USUBJID (subject ID), EINUM (AE ID),
EILLTN (LLT identifier), EIPTN (PT identifier), EISOCPN (soc identifier) and
EIGRDM (severity grade)}

\item{ref_grp}{(Optional) A reference group for comparisons. Defaults to the first group in \code{df_pat_grp}.}
}
\value{
A dataframe with all the info to build AE plots
}
\description{
Prepare a dataframe for creating AE plots
}
