% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indent_gtsummary_table.R
\name{indent_gtsummary_table}
\alias{indent_gtsummary_table}
\title{Text indentation for \code{gtsummary} tables}
\usage{
indent_gtsummary_table(gts_table, indent = 4)
}
\arguments{
\item{gts_table}{A \code{gtsummary} table object}

\item{indent}{A numerical value indicating how many space to indent text (default = 4). A value of 8 should be a good compromise.}
}
\value{
A \code{gtsummary} table object with indentation applied.
}
\description{
Text indentation for \code{gtsummary} tables
}
\examples{
tbl <- iris |>  dplyr::select(Species, Sepal.Length) |> RastaRocket::desc_var(
table_title = "test",
quali = "Species")
tbl_1 <- tbl |> indent_gtsummary_table(indent = 4)
tbl_2 <- tbl |> indent_gtsummary_table(indent = 8)
tbl_3 <- tbl |> indent_gtsummary_table(indent = 16)

}
