% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indent_table.R
\name{indent_table}
\alias{indent_table}
\title{Text indentation for \code{gtsummary} tables}
\usage{
indent_table(g_table, indent = 0)
}
\arguments{
\item{g_table}{A \code{gt} table object}

\item{indent}{A numeric value corresponding to value as pixels, specifying text indentation (default = 0). Value can be 0, 1 or 2 (0 ~ px(0); 1 ~ px(30); 2 ~ px(60))}
}
\value{
A \code{gt} table object with indentation applied.
}
\description{
Text indentation for \code{gtsummary} tables
}
\examples{
tbl_bis <- RastaRocket::desc_var(
iris,
table_title = "test",
quali = "Species")

tbl_bis |> indent_table(indent = 1)
tbl_bis |> indent_table(indent = 2)

}
