% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano Plot for Adverse Event Analysis}
\usage{
plot_volcano(
  df_soc_pt,
  df_pat_grp,
  df_pat_soc,
  colors_arm = c("#1b9e77", "#7570b3"),
  size = "nb_pat"
)
}
\arguments{
\item{df_soc_pt}{A data frame mapping system organ classes (SOC) to preferred terms (PT).
Must contain columns:
\itemize{
\item \code{soc} (character): System Organ Class.
\item \code{pt} (character): Preferred Term.
}}

\item{df_pat_grp}{A data frame with patient IDs and their assigned group.
Must contain columns:
\itemize{
\item \code{id_pat} (character): Patient ID.
\item \code{grp} (character): Group assignment (e.g., treatment arms).
}}

\item{df_pat_soc}{A data frame linking patients to reported SOCs.
Must contain columns:
\itemize{
\item \code{id_pat} (character): Patient ID.
\item \code{soc} (character): System Organ Class.
}}

\item{colors_arm}{A character vector of length two specifying the colors for the two patient groups in the plot.
Default is \code{c("#1b9e77", "#7570b3")}.}

\item{size}{A character string specifying the metric used for point sizes in the plot.
Options are:
\itemize{
\item \code{"nb_pat"}: Number of patients (default).
\item \code{"nb_ei"}: Number of adverse events.
}}
}
\value{
A ggplot2 object representing the volcano plot.
}
\description{
Generates a volcano plot to visualize the association between adverse events and patient groups.
}
\details{
The function first processes input data using \code{df_builder_ae()}, then calculates relevant statistics
such as risk difference (\code{RD}) and p-values. The volcano plot displays:
\itemize{
\item \code{RD} on the x-axis (risk difference between groups).
\item \code{-log10(p-value)} on the y-axis (significance level).
\item Point colors indicating which group has an increased risk.
\item Point sizes reflecting either the number of patients or events.
\item A horizontal dashed line at p = 0.05 for significance threshold.
}
}
\examples{
df_soc_pt <- data.frame(
  soc = c("Arrhythmia", "Myocardial Infarction", "Pneumonia", "Sepsis"),
  pt = c("Cardiac Disorders", "Cardiac Disorders", "Infections", "Infections")
)

df_pat_grp <- data.frame(id_pat = paste0("ID_", 1:10),
                         grp = c(rep("A", 5), rep("B", 5)))

df_pat_soc <- data.frame(id_pat = c("ID_1", "ID_1", "ID_2", "ID_4", "ID_9"),
                         soc = c("Arrhythmia", "Myocardial Infarction",
                         "Arrhythmia", "Pneumonia", "Pneumonia"))

plot_volcano(df_soc_pt, df_pat_grp, df_pat_soc)

}
