% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_table.R
\name{prepare_table}
\alias{prepare_table}
\title{Prepare a Data Frame for Summarization with Custom Missing Data Handling and Factor Ordering}
\usage{
prepare_table(
  data1,
  show_missing_data,
  drop_levels = TRUE,
  freq_relevel = FALSE
)
}
\arguments{
\item{data1}{A data frame containing the data to be prepared.}

\item{show_missing_data}{A boolean indicating whether missing data should be shown in the table.
If \code{NULL}, it will be automatically determined based on the presence of missing data in \code{data1}.
Defaults to \code{TRUE} if any missing values are detected.}

\item{drop_levels}{Boolean (default = TRUE). Drop unused levels.}

\item{freq_relevel}{Boolean  (default = FALSE). Reorder factors by frequency.}
}
\value{
A data frame that has been prepared based on the \code{show_missing_data} and \code{DM} arguments.
The function modifies the input data frame by applying labels, ordering factor variables,
and potentially dropping unused levels.
}
\description{
This function prepares a data frame for summarization by handling missing data
based on the \code{show_missing_data} argument and applying the specified data manipulation
(DM) option to factor variables. It provides flexibility for data cleaning and ordering
before summarizing with functions like \code{gtsummary}.
}
\details{
\itemize{
\item If \code{show_missing_data} is set to \code{TRUE} or \code{NULL} (default behavior), the function will apply labels
for missing data to the variables in the data.
\item If \code{show_missing_data} is set to \code{FALSE}, the function will display a warning if missing data is present.
\item The \code{DM} option defines the data manipulation to be applied to factor variables:
\itemize{
\item \code{"tout"}: Both order factor levels and drop unused levels.
\item \code{"tri"}: Only order factor levels.
\item \code{"remove"}: Drop unused factor levels without ordering.
}
}
}
\examples{
# Example usage with the iris dataset
prepare_table(iris, show_missing_data = TRUE)

}
