% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_qmd.R
\name{write_qmd}
\alias{write_qmd}
\title{Write a Quarto Markdown (.qmd) file}
\usage{
write_qmd(path, path_html, path_css)
}
\arguments{
\item{path}{Character string specifying the output file path for the .qmd file.}

\item{path_html}{Character string specifying the path to an HTML file to be included before the body of the document.}

\item{path_css}{Character string specifying the path to a CSS file for styling the document.}
}
\value{
None. The function writes a .qmd file to the specified \code{path}.
}
\description{
This function generates a Quarto Markdown (.qmd) file with predefined metadata and a sample table.
}
\details{
The function creates a Quarto Markdown file with metadata fields such as title, author, date, and format settings.
The HTML file specified in \code{path_html} is included before the body, and the CSS file specified in \code{path_css}
is used for styling. The generated document includes an example of a table with a caption.
}
